+++ man-roxygen/unname.R
@@ -1,0 +1,4 @@
+#' @param unname Should all chunks be unnamed before naming? Default is `FALSE`.
+#'   Turning this option on will cause all existing labels to be overwritten. In
+#'   contrast, with its default `unname = FALSE` only unlabelled chunks will be
+#'   named.