+++ tests/testthat/test-unname_chunks.R
@@ -1,19 +1,15 @@
-context("test-unname_all_chunks")
+context("test-unname_chunks")
 
-test_that("unname_all_chunks works in case is.null(chunk_name_prefix) == TRUE", {
+test_that("unname_chunks works in case is.null(chunk_name_prefix) == TRUE", {
   # check arg of tempdir
-  R_version <- paste(R.version$major,
-                     R.version$minor,
-                     sep = ".")
-
-  skip_if_not(R_version >= "3.5.0")
+  skip_if_not_r35()
   skip_if_not(rmarkdown::pandoc_available("1.12.3"))
 
   temp_file_path <- file.path(tempdir(check = TRUE), "example4.Rmd")
 
   file.copy(system.file("examples", "example4.Rmd", package = "namer"),
             temp_file_path)
-  unname_all_chunks(temp_file_path)
+  unname_chunks(temp_file_path)
 
   lines <- readLines(temp_file_path)
   chunk_info <- get_chunk_info(lines)
@@ -31,20 +27,16 @@ test_that("unname_all_chunks works in case is.null(chunk_name_prefix) == TRUE", 
   file.remove(paste0(basename, ".html"))
 })
 
-test_that("unname_all_chunks works in case is.null(chunk_name_prefix) == FALSE", {
+test_that("unname_chunks works in case is.null(chunk_name_prefix) == FALSE", {
   # check arg of tempdir
-  R_version <- paste(R.version$major,
-    R.version$minor,
-    sep = ".")
-
-  skip_if_not(R_version >= "3.5.0")
+  skip_if_not_r35()
   skip_if_not(rmarkdown::pandoc_available("1.12.3"))
 
   temp_file_path <- file.path(tempdir(check = TRUE), "example4.Rmd")
 
   file.copy(system.file("examples", "example4.Rmd", package = "namer"),
-    temp_file_path)
-  unname_all_chunks(temp_file_path,chunk_name_prefix='example4')
+            temp_file_path)
+  unname_chunks(temp_file_path,chunk_name_prefix='example4')
 
   lines <- readLines(temp_file_path)
   chunk_info <- get_chunk_info(lines)
@@ -63,20 +55,16 @@ test_that("unname_all_chunks works in case is.null(chunk_name_prefix) == FALSE",
   file.remove(paste0(basename, ".html"))
 })
 
-test_that("unname_all_chunks works in case chunk_name_prefix == 'setup' ", {
+test_that("unname_chunks works in case chunk_name_prefix == 'setup' ", {
   # check arg of tempdir
-  R_version <- paste(R.version$major,
-    R.version$minor,
-    sep = ".")
-
-  skip_if_not(R_version >= "3.5.0")
+  skip_if_not_r35()
   skip_if_not(rmarkdown::pandoc_available("1.12.3"))
 
   temp_file_path <- file.path(tempdir(check = TRUE), "example4.Rmd")
 
   file.copy(system.file("examples", "example4.Rmd", package = "namer"),
-    temp_file_path)
-  unname_all_chunks(temp_file_path,chunk_name_prefix='setup')
+            temp_file_path)
+  unname_chunks(temp_file_path,chunk_name_prefix='setup')
 
   lines <- readLines(temp_file_path)
   chunk_info <- get_chunk_info(lines)
@@ -96,3 +84,23 @@ test_that("unname_all_chunks works in case chunk_name_prefix == 'setup' ", {
 
   file.remove(paste0(basename, ".html"))
 })
+
+test_that("unname_all_chunks works but gives a warning",{
+  # check arg of tempdir
+  skip_if_not_r35()
+
+  temp_file_path <- file.path(tempdir(check = TRUE), "example4.Rmd")
+
+  file.copy(system.file("examples", "example4.Rmd", package = "namer"),
+            temp_file_path)
+  expect_warning(unname_all_chunks(temp_file_path),
+                 "please use")
+
+  lines <- readLines(temp_file_path)
+  chunk_info <- get_chunk_info(lines)
+
+  testthat::expect_identical(chunk_info$name[1],'setup')
+  testthat::expect_true(all(is.na(chunk_info$name[-1])))
+
+}
+)