+++ NEWS.md
@@ -1,5 +1,13 @@
 # namer (development version)
 
+* `name_chunks()` and `name_dir_chunks()` are now able to unname all chunks before naming them. This ensures a consistent naming for all chunks instead of just labelling unnamed chunks (@pat-s, #23).
+
+* new function `unname_dir_chunks()` that works in the same way as `name_dir_chunks()` (@pat-s, #23)
+
+* rename `unname_all_chunks()` to `unname_chunks()` to be consistent with `name_chunks()` (@pat-s, #23). `unname_all_chunks()` will be deprecated in a future version and currently gives a warning.
+
+* added some `cli` verbosity (@pat-s, #23)
+
 # namer 0.1.5
 
 * The tests using rmarkdown are now skipped if Pandoc 1.12.3 (minimal version for rmarkdown) is not available.